refactor: move PDF modules to obsidian-markdown + extract extractPdfText
pdf-engine.ts is PDF domain logic (pdfjs build swaps, font proxy wiring,
CMap path resolution) — it fails the utils/ admission bar of "generic with
zero domain knowledge." Move it to obsidian-markdown/ where its dependency
profile (node:module, node:path, unpdf, @napi-rs/canvas) is compatible with
the leaf-layer lint rules.

Create obsidian-markdown/pdf.ts with extractPdfText(Uint8Array → string) —
manages the pdfjs proxy lifecycle internally so callers pass bytes, get text.
Moves reconstructPdfMarkdown, groupIntoLines, buildHeadingLevels, roundFontSize
from asset-operations.ts (~120 lines).

asset-operations.ts's non-raw PDF branch now calls extractPdfText directly;
the raw/page-rendering branch keeps its own proxy for renderPdfPages.
