+++ src/vault-mcp/obsidian-markdown/pdf-engine.ts
